reddit_clone/requirements.txt
Django>=5.1,<5.2
graphene-django